requirements.txt
anyio==4.0.0
asgiref==3.7.2
async-timeout==4.0.3
attrs==23.1.0
autobahn==23.6.2
Automat==22.10.0
certifi==2023.7.22
cffi==1.15.1
channels==4.0.0
channels-redis==4.1.0
charset-normalizer==3.3.0
click==8.1.7
constantly==15.1.0
cryptography==41.0.3
daphne==4.0.0
defusedxml==0.8.0rc2
Django==4.2.3
django-cors-headers==4.1.0
django-taggit==4.0.0
django-taggit-serializer==0.1.7
django-templated-mail==1.1.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
drf-yasg==1.21.7
exceptiongroup==1.1.3
gunicorn==21.2.0
h11==0.14.0
httptools==0.6.0
hyperlink==21.0.0
idna==3.4
incremental==22.10.0
inflection==0.5.1
msgpack==1.0.5
oauthlib==3.2.2
packaging==23.1
Pillow==10.0.0
psycopg2-binary==2.9.6
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
PyJWT==2.8.0
pyOpenSSL==23.2.0
python-decouple==3.8
python-dotenv==1.0.0
python3-openid==3.2.0
pytz==2023.3.post1
PyYAML==6.0.1
redis==5.0.0
requests==2.31.0
requests-oauthlib==1.3.1
service-identity==23.1.0
six==1.16.0
sniffio==1.3.0
social-auth-app-django==5.3.0
social-auth-core==4.4.2
sqlparse==0.4.4
Twisted==23.8.0
txaio==23.1.1
typing_extensions==4.7.1
uritemplate==4.1.1
urllib3==2.0.6
uvicorn==0.23.2
uvloop==0.17.0
watchfiles==0.20.0
websocket-client==0.59.0
websockets==11.0.3
zope.interface==6.0
